Atualiza campos de salario de double para inteiro

+++ src/main/java/com/recrutaibackend/dto/ApplicationResponse.java
@@ -4,6 +4,6 @@ public record ApplicationResponse(
         int id,
         int candidateId,
         int vacancyId,
-        double expectedSalary
+        int expectedSalary
 ) {
 }
